Continue moving arelle.* package references into Brel.
Impacting arelle.XbrlConst mainly.  Also removal of renderer object
properties and variables containing the word 'arelle' as in
'arelleConcept', 'arelleValue', which seem to be holdovers from very
early code.  Chase down a few inappropriate uses of "_" as variable,
replacing with variables starting with "ignore".  Where this is not
practical, use pydev marker "UnusedVariable".  For reasons that are not
clear, the property 'clarkNotation' on QName objects shows up as
unbound, so some minor rewrites to deal with that.