First pass implementation of all deep-copy modelObjects in brel.
Currently, caching in the proxy object is turned off, but property
access recording is on.  Relationship sets are not well exercised.
In __init__, clarify the final debug logging message.
In Summary, move the detection of anchors into Utils.
In Utils, add typing hints where possible, and improve efficiency
of modelRelationshipsTransitiveFrom.